feat: only provides medium FAB and added widgetbook playground
