+++ lib/material_expressive.dart
@@ -3,4 +3,4 @@ library;
 
 export 'src/theme/m3e_floating_action_button_theme_data.dart';
 export 'src/theme/m3e_theme.dart';
-export 'src/widgets/buttons/m3e_floating_action_button.dart';
+export 'src/widgets/buttons/m3e_medium_floating_action_button.dart';
