+++ widgetbook/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:material_expressive/material_expressive.dart';
 import 'package:widgetbook/widgetbook.dart';
 import 'package:widgetbook_annotation/widgetbook_annotation.dart' as widgetbook;
 
@@ -16,10 +17,18 @@ class WidgetbookApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Widgetbook.material(
+    return Widgetbook(
       // The [directories] variable does not exist yet,
       // it will be generated in the next step
       directories: directories,
+      appBuilder: (context, child) {
+        return MaterialApp(
+          debugShowCheckedModeBanner: false,
+          theme: ThemeData.light().copyWith(extensions: [M3ETheme()]),
+          darkTheme: ThemeData.dark().copyWith(extensions: [M3ETheme()]),
+          home: Material(child: child),
+        );
+      },
     );
   }
 }
